perf: avoid needless string processing
