+++ src/components/MapEntity.tsx
@@ -1,201 +1,146 @@
-import { divIcon, LatLngExpression } from "leaflet";
-import React, { useMemo } from "react";
-import { renderToStaticMarkup } from "react-dom/server";
-import { Marker, Polyline } from "react-leaflet";
-import { planes } from "../dcs/aircraft";
+import * as maptalks from "maptalks";
+import ms from "milsymbol";
+import React, { useEffect } from "react";
 import {
   EntityTrackPing,
-  estimatedAltitudeRate,
   estimatedSpeed,
+  setTrackOptions,
   trackStore,
 } from "../stores/TrackStore";
 import { Entity } from "../types/entity";
-import { computeBRAA } from "../util";
-import { MapIcon } from "./MapIcon";
+import { getCardinal } from "../util";
+import { colorMode } from "./MapIcon";
 
-export function MapEntityTrail({ track }: { track: Array<EntityTrackPing> }) {
-  return (
-    <>
-      {track.map((ping, idx) => (
-        <Marker
-          key={ping.time}
-          position={ping.position}
-          icon={divIcon({
-            iconSize: [5, 5],
-            className: `bg-gray-100 bg-opacity-${100 - (idx * 10)}`,
-          })}
-          zIndexOffset={30}
-        />
-      ))}
-    </>
-  );
-}
+export const iconCache: Record<string, string> = {};
 
-function MapEntityInner({ entity }: { entity: Entity }) {
-  if (entity.types.includes("Air")) {
-    return <></>;
-  }
-
-  return (
-    <div className="flex flex-row absolute w-64">
-      <MapIcon
-        obj={entity}
-        className="relative bg-opacity-70"
-        size={16}
-      />
-    </div>
-  );
-}
-
-export function MapTrackedEntityInner(
-  { entity, active, track, hideLabel }: {
+export function EntityInfo(
+  { map, entity, track }: {
+    map: maptalks.Map;
     entity: Entity;
-    active: boolean;
     track: Array<EntityTrackPing>;
-    hideLabel: boolean;
   },
 ) {
-  const plane = planes[entity.name];
-  const speed = estimatedSpeed(track);
-
-  // Not visible to our radar
-  if (speed === -1 || speed < 25) {
-    return <></>;
-  }
-
-  const icon = useMemo(() => (
-    <MapIcon
-      obj={track.length < 5 ? "SPA-------" : entity}
-      className="relative bg-opacity-70"
-      size={16}
-    />
-  ), [entity.name, track.length]);
+  const trackOptions = trackStore((state) => state.trackOptions.get(entity.id));
 
   return (
-    <div className="flex flex-row z-40 w-80">
-      {icon}
-      {!hideLabel && (
-        <div
-          className="bg-gray-700 bg-opacity-40 flex flex-col absolute text-sm"
-          style={{ left: 24, top: -6 }}
-        >
-          {entity.types.includes("Air") && track.length >= 5 &&
-            (
-              <>
-                <div className="text-white">
-                  {entity.name}
-                  {entity.pilot && !entity.pilot.startsWith(entity.group)
-                    ? <>{" -"} {entity.pilot}</>
-                    : null}
-                </div>
-                <div className="flex flex-row gap-2">
-                  <div className="text-pink-300">
-                    {Math.floor(
-                      (entity.altitude * 3.28084) / 1000,
-                    )}
-                  </div>
-                  <>
-                    <div className="text-green-400">
-                      {Math.floor(estimatedSpeed(track))}
-                    </div>
-                    <div className="text-yellow-400">
-                      {Math.floor(estimatedAltitudeRate(track))}
-                    </div>
-                  </>
-                </div>
-              </>
-            )}
+    <div
+      className="m-2 absolute flex flex-col bg-gray-300 border border-gray-500 shadow select-none rounded-sm"
+    >
+      <div className="p-2 bg-gray-400 text-sm">
+        <b>{entity.group}</b>
+      </div>
+      <div className="p-2 flex flex-row">
+        <div className="flex flex-col pr-2">
+          <div>{entity.name}</div>
+          <div>{entity.pilot}</div>
           <div>
-            {active &&
-              (
-                <div className="flex flex-col">
-                  <span className="text-red-200">
-                    {plane && plane.natoName}
-                  </span>
-                  <span className="text-gray-100">
-                    {JSON.stringify(entity)}
-                  </span>
-                </div>
-              )}
+            Heading: {Math.round(entity.heading)}
+            {getCardinal(entity.heading)}
+          </div>
+          <div>Altitude: {Math.round(entity.altitude * 3.28084)}</div>
+          <div>GS: {Math.round(estimatedSpeed(track))}</div>
+        </div>
+        <div
+          className="flex flex-col border-l border-black px-2 gap-1 flex-grow"
+        >
+          <button
+            className="p-1 text-xs bg-blue-300 border border-blue-400"
+            onClick={() => {
+              map.animateTo({
+                center: [entity.longitude, entity.latitude],
+                zoom: 10,
+              }, {
+                duration: 250,
+                easing: "out",
+              });
+            }}
+          >
+            Snap
+          </button>
+          <div className="flex flex-col gap-1">
+            <div className="flex flex-row flex-grow">
+              <span className="text-yellow-600 pr-2 flex-grow">WR</span>
+              <input
+                className="w-16"
+                value={trackOptions?.warningRadius || ""}
+                onChange={(e) => {
+                  const val = parseInt(e.target.value);
+                  if (val !== NaN) {
+                    setTrackOptions(entity.id, {
+                      warningRadius: val,
+                    });
+                  }
+                }}
+              />
+            </div>
+            <div className="flex flex-row flex-grow">
+              <span className="text-red-600 pr-2 flex-grow">TR</span>
+              <input
+                className="w-16"
+                value={trackOptions?.threatRadius || ""}
+                onChange={(e) => {
+                  const val = parseInt(e.target.value);
+                  if (val !== NaN) {
+                    setTrackOptions(entity.id, {
+                      threatRadius: val,
+                    });
+                  }
+                }}
+              />
+            </div>
           </div>
         </div>
-      )}
+      </div>
     </div>
   );
 }
 
-export function MapEntity(
-  { entity, active, setActive, scale, hideLabel }: {
+export function MapSimpleEntity(
+  { map, entity, size, strokeWidth }: {
+    map: maptalks.Map;
     entity: Entity;
-    active: boolean;
-    setActive: () => void;
-    scale: number;
-    hideLabel: boolean;
+    size?: number;
+    strokeWidth?: number;
   },
 ) {
-  const position: LatLngExpression = [entity.latitude, entity.longitude];
-  const track = trackStore((state) => state.tracks.get(entity.id));
-
-  if (
-    entity.types.includes("Ground") ||
-    entity.types.length == 0
-  ) {
-    return <></>;
-  }
-
-  const icon = useMemo(() =>
-    divIcon({
-      html: renderToStaticMarkup(
-        track !== undefined
-          ? (
-            <MapTrackedEntityInner
-              entity={entity}
-              track={track}
-              active={active}
-              hideLabel={hideLabel}
-            />
-          )
-          : <MapEntityInner entity={entity} />,
-      ),
-      className: "",
-    }), [entity, active, track, hideLabel]);
-
-  const speed = track && estimatedSpeed(track);
-  const dirArrowEnd = speed && speed >= 25 && track && track.length >= 5 &&
-    computeBRAA(
-      position[0],
-      position[1],
-      entity.heading,
-      // knots -> meters per second -> 30 seconds
-      ((speed * 0.514444)) * 30,
-    );
-
-  return (
-    <>
-      {dirArrowEnd && entity.types.includes("Air") && (
-        <Polyline
-          positions={[
-            position,
-            dirArrowEnd,
-          ]}
-          pathOptions={{
-            color: entity.coalition !== "Allies" ? "#17c2f6" : "#ff8080",
-            weight: 1,
-          }}
-        />
-      )}
-      <Marker
-        position={position}
-        icon={icon}
-        eventHandlers={{
-          click: () => {
-            setActive();
+  useEffect(() => {
+    const iconLayer = map.getLayer("track-icons") as maptalks.VectorLayer;
+    let marker = iconLayer.getGeometryById(
+      entity.id,
+    ) as maptalks.Marker;
+    if (!marker) {
+      if (iconCache[entity.sidc] === undefined) {
+        iconCache[entity.sidc] = new ms.Symbol(entity.sidc, {
+          size: size || 16,
+          frame: true,
+          fill: false,
+          colorMode: colorMode,
+          strokeWidth: strokeWidth || 8,
+        }).toDataURL();
+      }
+      marker = new maptalks.Marker(
+        [entity.longitude, entity.latitude],
+        {
+          id: entity.id,
+          draggable: false,
+          visible: true,
+          editable: false,
+          symbol: {
+            markerFile: iconCache[entity.sidc],
+            markerDy: 10,
           },
-        }}
-        zIndexOffset={0}
-      />
-      {track && track.length >= 5 && speed && speed > 25 &&
-        <MapEntityTrail track={track} />}
-    </>
-  );
+        },
+      );
+      iconLayer.addGeometry(
+        marker,
+      );
+    } else {
+      marker.setCoordinates([
+        entity.longitude,
+        entity.latitude,
+      ]);
+    }
+  }, [entity]);
+
+  return <></>;
 }
